W1-20181226-17

+++ A_dbms/views/v_custom.py
@@ -123,14 +123,18 @@ def custom_del_ajax(request):
 
     custom_id = post_data['custom_id']
     custom_obj = models.Customes.objects.get(id=custom_id)
-
-    try:
-        custom_obj.delete()  # 删除评审会
-        msg = '%s，删除成功！' % custom_obj.name
-        response['message'] = msg
-    except Exception as e:
+    lending_custome_list = custom_obj.lending_custome.all()
+    if lending_custome_list:
         response['status'] = False
-        response['message'] = '客户删除失败:%s！' % str(e)
+        response['message'] = '客户已作为项目反担保人，无法删除！'
+    else:
+        try:
+            custom_obj.delete()  # 删除评审会
+            msg = '%s，删除成功！' % custom_obj.name
+            response['message'] = msg
+        except Exception as e:
+            response['status'] = False
+            response['message'] = '客户删除失败:%s！' % str(e)
 
     result = json.dumps(response, ensure_ascii=False)
     return HttpResponse(result)
