+++ A_dbms/views/v_warrant.py
@@ -94,19 +94,24 @@ def warrant_del_ajax(request):
 
     warrant_id = int(post_data['warrant_id'])
     warrant_obj = models.Warrants.objects.get(id=warrant_id)
+
     if warrant_obj.warrant_typ == 1:
         response['return_path'] = '/dbms/house/'
     elif warrant_obj.warrant_typ == 2:
         response['return_path'] = '/dbms/ground/'
 
-    print('warrant_obj:', warrant_obj)
-    try:
-        warrant_obj.delete()  # 删除评审会
-        msg = '%s，删除成功！' % warrant_obj.warrant_num
-        response['message'] = msg
-    except Exception as e:
+    lending_warrant_list = warrant_obj.lending_warrant.all()
+    if lending_warrant_list:
         response['status'] = False
-        response['message'] = '删除失败:%s！' % str(e)
+        response['message'] = '担保物已作为项目反担保，无法删除！'
+    else:
+        try:
+            warrant_obj.delete()  # 删除评审会
+            msg = '%s，删除成功！' % warrant_obj.warrant_num
+            response['message'] = msg
+        except Exception as e:
+            response['status'] = False
+            response['message'] = '删除失败:%s！' % str(e)
 
     result = json.dumps(response, ensure_ascii=False)
     return HttpResponse(result)
@@ -121,58 +126,97 @@ def warrant_edit_ajax(request):
     print('post_data:', post_data)
 
     warrant_id = int(post_data['warrant_id'])
-    print('warrant_id:', warrant_id)
-    print('type(warrant_id):', type(warrant_id))
     warrant_list = models.Warrants.objects.filter(id=warrant_id)
     warrant_obj = warrant_list[0]
     warrant_typ = warrant_obj.warrant_typ
 
-    if warrant_typ == 1:
-        print('warrant_typ == 1')
-        form_warrant_add_edit = forms.HouseAddForm(post_data)
-        if form_warrant_add_edit.is_valid():
-            warrant_add_data = form_warrant_add_edit.cleaned_data
-            try:
-                with transaction.atomic():
-                    warrant_list.update(
-                        warrant_num=warrant_add_data['warrant_num'])
+    form_warrant_edit = forms.WarrantEditForm(post_data)
 
-                    models.Houses.objects.filter(warrant=warrant_obj).update(
-                        house_locate=warrant_add_data['house_locate'],
-                        house_app=warrant_add_data['house_app'],
-                        house_area=warrant_add_data['house_area'])
-                response['message'] = '房产修改成功！！！，请继续创建产权证信息。'
-            except Exception as e:
+    if form_warrant_edit.is_valid():
+        warrant_edit_clean = form_warrant_edit.cleaned_data
+        print('warrant_edit_clean:', warrant_edit_clean)
+
+        if warrant_typ == 1:
+            print('warrant_typ == 1')
+            form_house_add_edit = forms.HouseAddEidtForm(post_data)
+            if form_house_add_edit.is_valid():
+                house_add_edit_clean = form_house_add_edit.cleaned_data
+                print('house_add_edit_clean:', house_add_edit_clean)
+                try:
+                    with transaction.atomic():
+                        warrant_list.update(
+                            warrant_num=warrant_edit_clean['warrant_num'])
+
+                        models.Houses.objects.filter(warrant=warrant_obj).update(
+                            house_locate=house_add_edit_clean['house_locate'],
+                            house_app=house_add_edit_clean['house_app'],
+                            house_area=house_add_edit_clean['house_area'])
+                    response['message'] = '房产修改成功！！！，请继续创建产权证信息。'
+                except Exception as e:
+                    response['status'] = False
+                    response['message'] = '房产修改失败：%s' % str(e)
+            else:
                 response['status'] = False
-                response['message'] = '房产修改失败：%s' % str(e)
-        else:
-            response['status'] = False
-            response['message'] = '表单信息有误！！！'
-            response['forme'] = form_warrant_add_edit.errors
+                response['message'] = '表单信息有误！！！'
+                response['forme'] = form_house_add_edit.errors
 
-    elif warrant_typ == 2:
-        print('warrant_typ == 2')
-        form_warrant_add_edit = forms.GroundAddForm(post_data)
-        if form_warrant_add_edit.is_valid():
-            warrant_add_data = form_warrant_add_edit.cleaned_data
-            print('warrant_add_data:', warrant_add_data)
-            try:
-                with transaction.atomic():
-                    warrant_list.update(
-                        warrant_num=warrant_add_data['warrant_num'])
+        elif warrant_typ == 2:
+            print('warrant_typ == 2')
+            form_ground_add_edit = forms.GroundAddEidtForm(post_data)
+            if form_ground_add_edit.is_valid():
+                ground_add_edit_clean = form_ground_add_edit.cleaned_data
+                print('ground_add_edit_clean:', ground_add_edit_clean)
+                try:
+                    with transaction.atomic():
+                        warrant_list.update(
+                            warrant_num=warrant_edit_clean['warrant_num'])
 
-                    models.Grounds.objects.filter(warrant=warrant_obj).update(
-                        ground_locate=warrant_add_data['ground_locate'],
-                        ground_app=warrant_add_data['ground_app'],
-                        ground_area=warrant_add_data['ground_area'])
-                    response['message'] = '土地创建成功！！！，请继续创建产权证信息。'
-            except Exception as e:
+                        models.Grounds.objects.filter(warrant=warrant_obj).update(
+                            ground_locate=ground_add_edit_clean['ground_locate'],
+                            ground_app=ground_add_edit_clean['ground_app'],
+                            ground_area=ground_add_edit_clean['ground_area'])
+                        response['message'] = '土地创建成功！！！，请继续创建产权证信息。'
+                except Exception as e:
+                    response['status'] = False
+                    response['message'] = '土地创建失败：%s' % str(e)
+            else:
                 response['status'] = False
-                response['message'] = '土地创建失败：%s' % str(e)
-        else:
+                response['message'] = '表单信息有误！！！'
+                response['forme'] = form_ground_add_edit.errors
+    else:
+        response['status'] = False
+        response['message'] = '表单信息有误！！！'
+        response['forme'] = form_warrant_edit.errors
+    result = json.dumps(response, ensure_ascii=False)
+    return HttpResponse(result)
+
+
+# -----------------------产权证添加-------------------------#
+def owership_add_ajax(request):
+    print(__file__, '---->def owership_add_ajax')
+    response = {'status': True, 'message': None, 'forme': None, }
+    post_data_str = request.POST.get('postDataStr')
+    post_data = json.loads(post_data_str)
+    print('post_data:', post_data)
+    warrant_id = post_data['warrant_id']
+    warrant_obj = models.Warrants.objects.get(id=warrant_id)
+    form_owership_add_edit = forms.OwerShipEditForm(post_data)
+    if form_owership_add_edit.is_valid():
+        owership_add_clean = form_owership_add_edit.cleaned_data
+        print('warrant_add_clean:', owership_add_clean)
+        try:
+            owership_obj = models.Ownership.objects.create(
+                warrant=warrant_obj,
+                ownership_num=owership_add_clean['ownership_num'],
+                owner=owership_add_clean['owner'])
+            response['message'] = '产权证信息创建成功！！！'
+        except Exception as e:
             response['status'] = False
-            response['message'] = '表单信息有误！！！'
-            response['forme'] = form_warrant_add_edit.errors
+            response['message'] = '产权证信息创建失败：%s' % str(e)
+    else:
+        response['status'] = False
+        response['message'] = '表单信息有误！！！'
+        response['forme'] = form_owership_add_edit.errors
     result = json.dumps(response, ensure_ascii=False)
     return HttpResponse(result)
 
@@ -202,27 +246,6 @@ def house(request, *args, **kwargs):  # 房产列表
 
     return render(request,
                   'dbms/warrant/house.html',
-                  locals())
-
-
-# ---------------------house_scan房产预览------------------------#
-def house_scan(request, house_id):  # house_scan房产预览
-    print(__file__, '---->def house_scan')
-    warrant_obj = models.Houses.objects.get(id=house_id)
-    edit_warrant = '修改房产'
-    del_warrant = '删除房产'
-    warrant_typ_n = warrant_obj.warrant.warrant_typ
-    form_date = {
-        'warrant_num': warrant_obj.warrant,
-        'house_locate': warrant_obj.house_locate,
-        'house_app': warrant_obj.house_app,
-        'house_area': warrant_obj.house_area}
-
-    form_warrant_edit = forms.WarrantEditForm(form_date)
-    form_house_add_edit = forms.HouseAddEidtForm(form_date)
-
-    return render(request,
-                  'dbms/warrant/house-scan.html',
                   locals())
 
 
@@ -279,18 +302,30 @@ def warrant(request, *args, **kwargs):  # 房产列表
                   locals())
 
 
-# ---------------------hwarrant_scan权证预览------------------------#
+# ---------------------warrant_scan权证预览------------------------#
 def warrant_scan(request, warrant_id):  # house_scan房产预览
-    print(__file__, '---->def warrant_scan')
-    house_obj = models.Warrants.objects.get(id=warrant_id)
+    print(__file__, '---->def house_scan')
+    warrant_obj = models.Warrants.objects.get(id=warrant_id)
+    edit_warrant = '修改房产'
+    del_warrant = '删除房产'
+    warrant_typ_n = warrant_obj.warrant_typ
+    form_warrant_edit_date = {'warrant_num': warrant_obj.warrant_num}
+    warrant_typ = warrant_obj.warrant_typ
+    if warrant_typ == 1:
+        form_date = {
+            'house_locate': warrant_obj.house_warrant.house_locate,
+            'house_app': warrant_obj.house_warrant.house_app,
+            'house_area': warrant_obj.house_warrant.house_area}
+        form_house_add_edit = forms.HouseAddEidtForm(form_date)
+    elif warrant_typ == 2:
+        form_date = {
+            'ground_locate': warrant_obj.ground_warrant.ground_locate,
+            'ground_app': warrant_obj.ground_warrant.ground_app,
+            'ground_area': warrant_obj.ground_warrant.ground_area}
+        form_ground_add_edit = forms.GroundAddEidtForm(form_date)
 
-    form_date = {
-        'house_locate': house_obj.house_locate,
-        'house_app': house_obj.house_app,
-        'house_area': house_obj.house_area}
+    form_warrant_edit = forms.WarrantEditForm(form_warrant_edit_date)
+    form_owership_add_edit = forms.OwerShipEditForm()
 
-    form_house_add_edit = forms.HouseAddForm(form_date)
-
-    return render(request,
-                  'dbms/warrant/house-scan.html',
-                  locals())
+    print('form_owership_add_edit:', form_owership_add_edit)
+    return render(request, 'dbms/warrant/warrant-scan.html', locals())
